Implement immunity by locus
The option `whole_gene_immune` controls whether a host gains immunity towards a gene if the host has seen all the alleles. If `true`, then a host gains immunity towards a gene only if the host has seen all the alleles. If `false`, then a host gains a gradual increase in immunity as the host sees more alleles in a gene. In the latter case, instead of using the rejection sampling which could slow down the simulation substantially, I operated the switching rate, immunity incrementation, and immunity decrementation at the locus level (instead of the gene level). In other words:
1. The `switching_rate` is defined at the locus level (i.e. the allele switching rate = `switching_rate` / `n_loci`; e.g. if `switching_rate` = 0.2 and if  `n_loci` = 2, then the allele switching rate = 0.2 / 2 = 0.1).
2. A locus expression index is created for each host (e.g. its `expression_index_locus = 1` if the expressing allele is located at the first locus of the gene, and the `expression_index_locus = 2` if the expressing allele is located at the second locus of the gene). This `expression_index_locus` is used to operate the switching event, immunity incrementation, and immunity decrementation at the locus level.