Add Rscript_Create_LDPairwiseMatrix.R
The R script "Rscript_Create_LDPairwiseMatrix.R" could be used to create a pairwise linkage disequilibrium (LD) matrix.This matrix should provide the coefficient of LD between each pair of SNPs. If two loci are not coinherited at all (they are independent) then the value will be 0.0. If two loci are in total disequilibrium then value would be 1.0.